functions should always stay close to where they are used
