+++ pkg/apiCalls/prepare_test.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/xaaha/hulak/pkg/utils"
+	"github.com/xaaha/hulak/pkg/yamlParser"
 )
 
 func TestFullUrl(t *testing.T) {
@@ -126,7 +126,7 @@ func TestEncodeXwwwFormUrlBody(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			reader, err := utils.EncodeXwwwFormUrlBody(tt.input)
+			reader, err := yamlParser.EncodeXwwwFormUrlBody(tt.input)
 
 			// Check if an error is expected
 			if tt.expectError {
@@ -203,7 +203,7 @@ func TestEncodeFormData(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			payload, contentType, err := utils.EncodeFormData(tt.input)
+			payload, contentType, err := yamlParser.EncodeFormData(tt.input)
 
 			// Check if an error is expected
 			if tt.expectError {
@@ -287,7 +287,7 @@ func TestEncodeGraphQlBody(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			payload, err := utils.EncodeGraphQlBody(tt.query, tt.variables)
+			payload, err := yamlParser.EncodeGraphQlBody(tt.query, tt.variables)
 
 			// Check if an error is expected
 			if tt.expectError {
