--- pkg/utils/apiUtils.go
@@ -1,69 +1,0 @@
-package utils
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io"
-	"mime/multipart"
-	"net/url"
-	"strings"
-)
-
-// Encodes key-value pairs as "application/x-www-form-urlencoded" data.
-// Returns an io.Reader containing the encoded data, or an error if the input is empty.
-func EncodeXwwwFormUrlBody(keyValue map[string]string) (io.Reader, error) {
-	// Initialize form data
-	formData := url.Values{}
-
-	// Populate form data, using Set to overwrite duplicate keys if any
-	for key, val := range keyValue {
-		if key != "" && val != "" {
-			formData.Set(key, val)
-		}
-	}
-
-	// Return an error if no valid key-value pairs were found
-	if len(formData) == 0 {
-		return nil, fmt.Errorf("no valid key-value pairs to encode")
-	}
-
-	// Encode form data to "x-www-form-urlencoded" format
-	return strings.NewReader(formData.Encode()), nil
-}
-
-// Encodes multipart/form-data other than x-www-form-urlencoded,
-// Returns the payload, Content-Type for the headers and error
-func EncodeFormData(keyValue map[string]string) (io.Reader, string, error) {
-	if len(keyValue) == 0 {
-		return nil, "", fmt.Errorf("no key-value pairs to encode")
-	}
-
-	payload := &bytes.Buffer{}
-	writer := multipart.NewWriter(payload)
-	defer writer.Close() // Ensure writer is closed
-
-	for key, val := range keyValue {
-		if key != "" && val != "" {
-			if err := writer.WriteField(key, val); err != nil {
-				return nil, "", err
-			}
-		}
-	}
-
-	// Return the payload and the content type for the header
-	return payload, writer.FormDataContentType(), nil
-}
-
-// accepts query string and variables map[string]interface, then returns the payload
-func EncodeGraphQlBody(query string, variables map[string]interface{}) (io.Reader, error) {
-	payload := map[string]interface{}{
-		"query":     query,
-		"variables": variables,
-	}
-	jsonData, err := json.Marshal(payload)
-	if err != nil {
-		return nil, err
-	}
-	return bytes.NewReader(jsonData), nil
-}
